[refactor]: 컴포넌트 폴더 내 페이지별 폴더 생성 및 파일 이동
