+++ src/components/Home/ProductsByTheme.js
@@ -1,6 +1,10 @@
+/** 테마id에 따른 상품 목록을 표현하기 위한 컴포넌트
+ *  - ProductItem 컴포넌트로 해당 테마에 포함된 각 상품 표현
+ */
+
 import { useState, useEffect } from "react";
 import styled from "styled-components";
-import { getThemeProducts } from "../apis/ShoppingAPI";
+import { getThemeProducts } from "../../apis/ShoppingAPI";
 import ProductItem from "./ProductItem";
 
 const Box = styled.div`
